ipv6: cleanup pass before review
