+++ internalshared/configutil/telemetry.go
@@ -36,7 +36,6 @@ const (
 )
 
 // Telemetry is the telemetry configuration for the server.
-// NOTE: When adding new address or URL fields be sure to update normalizeTelemetryAddresses().
 type Telemetry struct {
 	FoundKeys    []string     `hcl:",decodedFields"`
 	UnusedKeys   UnusedKeyMap `hcl:",unusedKeyPositions"`
@@ -194,7 +193,9 @@ func parseTelemetry(result *SharedConfig, list *ast.ObjectList) error {
 		return multierror.Prefix(err, "telemetry:")
 	}
 
-	// Make sure addresses conform to RFC-5952
+	// Make sure addresses conform to RFC-5942 §4. If you've added new fields that
+	// are an address or URL be sure to update normalizeTelemetryAddresses().
+	// See: https://rfc-editor.org/rfc/rfc5952.html
 	normalizeTelemetryAddresses(result.Telemetry)
 
 	if result.Telemetry.PrometheusRetentionTimeRaw != nil {
@@ -246,8 +247,8 @@ func parseTelemetry(result *SharedConfig, list *ast.ObjectList) error {
 	return nil
 }
 
-// normalizeTelemetryAddresses takes a reference to Telemetry and ensures that
-// any address and URL configuration options that are IPv6 conform to RFC-5952.
+// normalizeTelemetryAddresses ensures that any telemetry configuration that can
+// be a URL, IP Address, or host:port address is conformant with RFC-5942 §4
 // See: https://rfc-editor.org/rfc/rfc5952.html
 func normalizeTelemetryAddresses(in *Telemetry) {
 	if in == nil {
